requirements.txt
Click>=7.0
gitpython>=3.1.18
mcstasscript
